[add] /about route to print user info + a little refactoring
