+++ config.py
@@ -18,3 +18,4 @@ class Config:
     PROTOCOL_DEFAULT = config['SERVER']['SC_SERVER_PROTOCOL_DEFAULT']
     HOST_DEFAULT = config['SERVER']['SC_SERVER_HOST_DEFAULT']
     PORT_DEFAULT = config['SERVER']['SC_SERVER_PORT_DEFAULT']
+    MAX_SESSION_SIZE = 4093
