+++ service/agents/ostis.py
@@ -3,7 +3,6 @@ from threading import Event
 import re
 
 import sc_client.client as client
-from sc_client.client import generate_by_template
 from ..exceptions import ScServerError
 from sc_client.client import is_connected
 from sc_client.models import (
@@ -19,9 +18,7 @@ from sc_client.constants.common import ScEventType
 from sc_client.constants import sc_types
 from sc_kpm.utils.common_utils import (
     generate_node, 
-    generate_role_relation, 
-    get_link_content_data, 
-    get_element_system_identifier, 
+    generate_role_relation,
     )
 from sc_kpm import ScKeynodes
 
@@ -31,6 +28,12 @@ from service.agents.abstract.reg_agent import RegAgent, RegStatus
 from service.agents.abstract.user_request_agent import RequestAgent, RequestStatus
 from service.agents.abstract.directory_agent import DirectoryAgent, DirectoryStatus
 from service.exceptions import AgentError, ParseDataError
+from service.utils.ostis_utils import(
+    create_link,
+    get_node,
+    set_gender_content,
+    set_birthdate_content
+)
 from config import Config
 
 payload = None
@@ -43,58 +46,7 @@ gender_dict = {
 
 class result(Enum):
     SUCCESS = 0
-    FAILURE = 1
-
-def create_link(client, content: str):
-    construction = ScConstruction()
-    to_find_content = ScLinkContent(content, ScLinkContentType.STRING)
-    construction.create_link(sc_types.LINK_CONST, to_find_content)
-    link = client.generate_elements(construction)
-    return link[0]
-
-def get_node(client) -> ScAddr:
-    construction = ScConstruction()
-    construction.create_node(sc_types.NODE_CONST)
-    main_node: ScAddr = client.generate_elements(construction)[0]
-    return main_node
-
-def get_main_idtf(node: ScAddr) -> str:
-    template = ScTemplate()
-    template.quintuple(
-        node,
-        sc_types.EDGE_D_COMMON_VAR,
-        '_value',
-        sc_types.EDGE_ACCESS_VAR_POS_PERM,
-        ScKeynodes['nrel_main_idtf']
-        )
-    template_result = client.template_search(template)
-    value = ''
-    if len(template_result):
-        value = client.get_link_content(template_result[0].get('_value'))[0].data
-    return value
-
-def set_gender_content(gender) -> ScAddr:
-    if gender == "male":
-        return ScKeynodes['concept_man']
-    if gender == "female":
-        return ScKeynodes['concept_woman']
-    else:
-        raise ParseDataError(666, "Failed to parse args")
-
-def set_birthdate_content(client, birthdate):
-    pattern = r'(\d{2})\.(\d{2})\.(\d{4})'
-    match = re.match(pattern, birthdate)
-    if match:
-        day, month, year = map(int, match.groups())
-        print(day, month, year)
-        _day_lnk = create_link(client, day)
-        _month_lnk = create_link(client, month)
-        _year_lnk = create_link(client, year)
-
-        return _day_lnk, _month_lnk, _year_lnk
-    else:
-        raise ParseDataError(666, "Failed to parse args") 
-
+    FAILURE = 1 
 
 def call_back(src: ScAddr, connector: ScAddr, trg: ScAddr) -> Enum:
     global payload
@@ -273,8 +225,8 @@ def call_back_directory(src: ScAddr, connector: ScAddr, trg: ScAddr) -> Enum:
             content_data = client.get_link_content(_content_link)[0].data
             content_list.append(
                 DirectoryResponse(
-                    title=title_data[:300],
-                    content=content_data[:300] + "...")
+                    title=title_data,
+                    content=content_data)
                 )
         payload = {"message": content_list}
     elif trg.value == unsucc_node.value or trg.value == node_err.value:
